feat: update follow-list (#193) 버튼들 page로 변경
