+++ components/Domain/FollowList/Follower/index.tsx
@@ -1,14 +1,12 @@
 import { Body3, Button3 } from '@/components/UI';
 import S from './style';
 import { useRouter } from 'next/router';
-import FollowAlert from '../FollowAlert';
-import { useState } from 'react';
-import Toast from '@/components/Toast';
 import SearchBar from '@/components/SearchBar';
 import { Dispatch, SetStateAction, useEffect } from 'react';
 import { followListType } from '@/pages/follow-list/[...UserId]';
 
 interface followerProps {
+  setAlertOpen: Dispatch<SetStateAction<Boolean>>;
   keyword: string;
   setKeyword: Dispatch<SetStateAction<string>>;
   followerList?: followListType[];
@@ -17,6 +15,7 @@ interface followerProps {
 }
 
 export default function Follower({
+  setAlertOpen,
   keyword,
   setKeyword,
   followerList,
@@ -24,23 +23,11 @@ export default function Follower({
   showingId,
 }: followerProps) {
   const router = useRouter();
-  const [alertOpen, setAlertOpen] = useState<Boolean>(false);
-  const [toastOpen, setToastOpen] = useState<Boolean>(false);
 
   console.log(followerList);
 
   const onClickDelete = () => {
     setAlertOpen(true);
-  };
-
-  const onClickYesButton = () => {
-    console.log('팔로우');
-    setAlertOpen(false);
-    setToastOpen(true);
-  };
-
-  const onClickNoButton = () => {
-    setAlertOpen(false);
   };
 
   const onClickFollow = async (status: Boolean, id: number) => {
@@ -51,7 +38,6 @@ export default function Follower({
 
   return (
     <>
-      <S.Background isOpen={alertOpen} onClick={() => setAlertOpen(false)} />
       <S.Wrap>
         <SearchBar placeholder="검색" letter={keyword} setLetter={setKeyword} />
       </S.Wrap>
@@ -59,41 +45,35 @@ export default function Follower({
         {followerList &&
           followerList.map((item, index) => {
             return (
-              <S.FollowBlockLayout key={index}>
-                <img src={item.userImage} alt="" />
-                <Body3 state="emphasis" className="name">
-                  {item.userName}
-                </Body3>
-                {localUserId === showingId ? (
-                  <Button3 className="delete" onClick={() => onClickDelete()}>
-                    삭제
-                  </Button3>
-                ) : item.isFollow ? (
-                  <Button3
-                    className="unfollow"
-                    onClick={() => onClickFollow(item.isFollow, item.userId)}
-                  >
-                    팔로잉
-                  </Button3>
-                ) : (
-                  <Button3
-                    className="following"
-                    onClick={() => onClickFollow(item.isFollow, item.userId)}
-                  >
-                    팔로우
-                  </Button3>
-                )}
-              </S.FollowBlockLayout>
+              <>
+                <S.FollowBlockLayout key={index}>
+                  <img src={item.userImage} alt="" />
+                  <Body3 state="emphasis" className="name">
+                    {item.userName}
+                  </Body3>
+                  {localUserId === showingId ? (
+                    <Button3 className="delete" onClick={() => onClickDelete()}>
+                      삭제
+                    </Button3>
+                  ) : item.isFollow ? (
+                    <Button3
+                      className="unfollow"
+                      onClick={() => onClickFollow(item.isFollow, item.userId)}
+                    >
+                      팔로잉
+                    </Button3>
+                  ) : (
+                    <Button3
+                      className="following"
+                      onClick={() => onClickFollow(item.isFollow, item.userId)}
+                    >
+                      팔로우
+                    </Button3>
+                  )}
+                </S.FollowBlockLayout>
+              </>
             );
           })}
-
-        {alertOpen && (
-          <FollowAlert
-            onClickYesButton={onClickYesButton}
-            onClickNoButton={onClickNoButton}
-          />
-        )}
-        {toastOpen && <Toast text="삭제되었습니다." />}
       </S.Layout>
     </>
   );
